+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Don't blow up when a Rodauth request is made using an unsupported HTTP verb (@janko)
+
 ## 0.7.0 (2020-11-27)
 
 * Add `#rails_controller_eval` method for running code in context of a controller instance (@janko)